Lab6/requirements.txt
Flask==3.0.2
PyJWT==2.8.0
python-dotenv==1.0.1